# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Mpango ERP API
  description: |
    Multi-tenant ERP system for African wholesale-retail operations.

    ## Authentication
    All endpoints except `/auth/login` require Bearer token authentication.
    The JWT token contains `tenant_id`, `tenant_schema`, and `user_id` claims.

    ## Multi-Tenancy
    - Login requires `tenant_code` to identify the wholesaler
    - All subsequent requests use tenant info from JWT claims
    - Data is isolated per tenant schema

    ## Authorization
    Endpoints are protected by RBAC permissions as defined in RBAC Matrix (MVP).
  version: 1.0.0
  contact:
    name: Mpango ERP Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mpango.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: List users with pagination
      description: Requires `users:read` permission
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [users]
      operationId: createUser
      summary: Create a new user
      description: Requires `users:create` permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{user_id}:
    get:
      tags: [users]
      operationId: getUser
      summary: Get user by ID
      description: Requires `users:read` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [users]
      operationId: updateUser
      summary: Update user
      description: Requires `users:update` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [users]
      operationId: deleteUser
      summary: Soft delete user (deactivate)
      description: Requires `users:deactivate` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{user_id}/roles:
    put:
      tags: [users]
      operationId: assignUserRoles
      summary: Assign roles to user
      description: Requires `roles:assign` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest'
      responses:
        '200':
          description: Roles assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # ROLE ENDPOINTS
  # ============================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing claims:
        - user_id: UUID
        - tenant_id: UUID
        - tenant_schema: string (e.g., t_<uuid>)
        - exp: expiration timestamp

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    UserIdParam:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    Pagination:
      type: object
      required: [page, size, total, pages]
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        meta:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Input validation failed
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time

    # ========== Payment ==========
...
    UserCreateRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: newuser@acme.com
        password:
          type: string
          format: password
          minLength: 8
        full_name:
          type: string
          maxLength: 100
          example: John Doe
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
          maxLength: 100
        is_active:
          type: boolean
    User:
      type: object
      required: [id, email, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/User'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    AssignRolesRequest:
      type: object
      required: [role_ids]
      properties:
        role_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1

    # ========== Role ==========
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: admin
        description:
          type: string
...